nvim: add formatter
Format the current buffer with the provided formatters.

Multiple formatters can be used at the same time, and are applied in the
provided order. The same formatter can be used multiple times.

Available formatters (for now):
- *count*: count the number of occurrences of each unique line,
and sort the count in descending order. Equivalent to
`% !sort | uniq -c | sort -nr`.
- *selinux*: filter SELinux violations logs, removing irrelevant
  content.
- *sortuniq*: sort lines, uniquely. Equivalent to `% !sort -u`.

Example usage:
- Show available formatters: `:Format`
- Format with *selinux*, followed by *count*: `:Format selinux count`